Move matches_platform to 'common' module to prevent import hell
Putting it to rule module was a problem if you don't have docker.
Addressing:
_____ ERROR collecting
tests/unit/ssg_test_suite/test_matches_platform.py ______
ImportError while importing test module
'/home/jenkins/workspace/scap-security-guide-pull-requests/label/fedora/tests/unit/ssg_test_suite/test_matches_platform.py'.
Hint: make sure your test modules/packages have valid Python names.
Traceback:
../../tests/unit/ssg_test_suite/test_matches_platform.py:3: in <module>
    from ssg_test_suite import rule
../../tests/ssg_test_suite/rule.py:12: in <module>
    import ssg_test_suite.oscap as oscap
../../tests/ssg_test_suite/oscap.py:13: in <module>
    from ssg_test_suite import test_env
../../tests/ssg_test_suite/test_env.py:8: in <module>
    import docker
E   ModuleNotFoundError: No module named 'docker'
